+++ frontend/src/pages/Home.vue
@@ -84,9 +84,9 @@
         </a>
         <p class="small">
             <br>
-            Learn more about the security of <em>EVE Online SSO</em> in this
-            <a href="https://www.eveonline.com/article/eve-online-sso-and-what-you-need-to-know/"
-               target="_blank" rel="noopener noreferrer">dev-blog</a> article.
+            Learn more about
+            <a href="https://support.eveonline.com/hc/en-us/articles/205381192"
+               target="_blank" rel="noopener noreferrer">EVE Online Single Sign On</a>.
         </p>
         <span v-if="markdownLoginText">
             <br>